dashboard section refactored
