+++ src/helpers/sendVerificationEmail.js
@@ -10,7 +10,7 @@ export async function sendVerificationEmail(
         await resend.emails.send({
             from: 'onboarding@resend.dev',
             to:email,
-            subject: 'True Feedback Message Verification Code',
+            subject: 'Feedify Feedback Message Verification Code',
             react: VerificationEmail({ username, otp: verifyCode }),
         });
         return { success: true, message: 'Verification email sent successfully.' };
